samples/onnx_accelerator_sample1/requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
aws-cdk-lib>=2.1.0, <3.0.0
boto3==1.26.72
cdk-nag==2.22.28
constructs>=10.0.5, <11.0.0
